[UPDATE] move some methods to parent class
